Nag add ko ug enter function using onkeypress para sa search bar and gi-isa na nako ang function for search

+++ database/migrations/2018_05_20_161801_update_start_and_end_dates.php
@@ -16,14 +16,16 @@ class UpdateStartAndEndDates extends Migration
         Schema::table('semesters', function (Blueprint $table) {
             if (Schema::hasColumn('semesters', 'start_date')) {
                 $table->dropColumn('start_date');
+            } else {
+                $table->date('start_date')->nullable();   
             }
 
             if (Schema::hasColumn('semesters', 'end_date')) {
                 $table->dropColumn('end_date');
+            } else {
+                $table->date('end_date')->nullable();
             }
             
-            $table->date('start_date')->nullable();
-            $table->date('end_date')->nullable();
         });
 
     }
@@ -36,8 +38,13 @@ class UpdateStartAndEndDates extends Migration
     public function down()
     {
         Schema::table('semesters', function (Blueprint $table) {
-            $table->dropColumn('start_date');
-            $table->dropColumn('end_date');
+            if (Schema::hasColumn('semesters', 'start_date')) {
+                $table->dropColumn('start_date');
+            }
+
+            if (Schema::hasColumn('semesters', 'end_date')) {
+                $table->dropColumn('end_date');
+            }
 
             $table->string('start_date');
             $table->string('end_date');
